moving all functions
