+++ cartoframes/data/dataframe_dataset.py
@@ -7,6 +7,20 @@ class DataFrameDataset(DatasetBase):
         self._state = DatasetBase.STATE_LOCAL
 
         self._df = data
+        _save_index_as_column(self._df)
+
+    @property
+    def df(self):
+        """Dataset DataFrame"""
+        return self._df
+
+    def get_geodataframe(self):
+        """Converts DataFrame into GeoDataFrame if possible"""
+        gdf = compute_geodataframe(self)
+        if not gdf.empty:
+            self._df = gdf
+
+        return self._df
 
     def download(self):
         self._is_ready_for_dowload_validation()
@@ -14,19 +28,23 @@ class DataFrameDataset(DatasetBase):
     def upload(self, if_exists, with_lnglat):
         self._is_ready_for_upload_validation()
 
+        normalized_column_names = _normalize_column_names(self._df)
+
         if if_exists == DatasetBase.REPLACE or not self.exists():
-            self._create_table(with_lnglat)
-            # if if_exists != DatasetBase.APPEND:
-            #     self._is_saved_in_carto = True
+            self._create_table(normalized_column_names, with_lnglat)
         elif if_exists == DatasetBase.FAIL:
             raise self._already_exists_error()
 
-        self._copyfrom(with_lnglat)
+        self._copyfrom(normalized_column_names, with_lnglat)
 
-    def _copyfrom(self, with_lnglat):
+    def compute_geom_type(self):
+        """Compute the geometry type from the data"""
+        return self._get_geom_type()
+
+    def _copyfrom(self, normalized_column_names, with_lnglat):
         geom_col = _get_geom_col_name(self._df)
         enc_type = _detect_encoding_type(self._df, geom_col)
-        columns = ','.join(norm for norm, orig in _normalize_column_names(self._df))
+        columns = ','.join(norm for norm, orig in normalized_column_names)
 
         query = """COPY {table_name}({columns},the_geom) FROM stdin WITH (FORMAT csv, DELIMITER '|');""".format(
             table_name=self._table_name,
@@ -40,6 +58,43 @@ class DataFrameDataset(DatasetBase):
             enc_type)
 
         self._client.upload(query, data)
+
+    def _create_table(self, normalized_column_names, with_lnglat=None):
+        query = '''BEGIN; {drop}; {create}; {cartodbfy}; COMMIT;'''.format(
+            drop=self._drop_table_query(),
+            create=self._create_table_query(normalized_column_names, with_lnglat),
+            cartodbfy=self._cartodbfy_query())
+
+        try:
+            self._client.execute_long_running_query(query)
+        except CartoRateLimitException as err:
+            raise err
+        except CartoException as err:
+            raise CartoException('Cannot create table: {}.'.format(err))
+
+    def _create_table_query(self, normalized_column_names, with_lnglat=None):
+        if with_lnglat is None:
+            geom_type = _get_geom_col_type(self._df)
+        else:
+            geom_type = 'Point'
+
+        col = ('{col} {ctype}')
+        cols = ', '.join(col.format(col=norm,
+                                    ctype=_dtypes2pg(self._df.dtypes[orig]))
+                         for norm, orig in normalized_column_names)
+
+        if geom_type:
+            cols += ', {geom_colname} geometry({geom_type}, 4326)'.format(geom_colname='the_geom', geom_type=geom_type)
+
+        create_query = '''CREATE TABLE {table_name} ({cols})'''.format(table_name=self._table_name, cols=cols)
+        return create_query
+
+    def _get_geom_type(self):
+        """Compute geom type of the local dataframe"""
+        if not self._df.empty and 'geometry' in self._df and len(self._df.geometry) > 0:
+            geometry = _first_value(self._df.geometry)
+            if geometry and geometry.geom_type:
+                return self._map_geom_type(geometry.geom_type)
 
 
 def _rows(df, cols, with_lnglat, geom_col, enc_type):
@@ -119,3 +174,45 @@ def _detect_encoding_type(df, geom_col):
         if first_geom:
             return detect_encoding_type(first_geom)
     return ''
+
+
+def _save_index_as_column(df):
+    index_name = df.index.name
+    if index_name is not None:
+        if index_name not in df.columns:
+            df.reset_index(inplace=True)
+            df.set_index(index_name, drop=False, inplace=True)
+
+
+def _dtypes2pg(dtype):
+    """Returns equivalent PostgreSQL type for input `dtype`"""
+    mapping = {
+        'float64': 'numeric',
+        'int64': 'integer',
+        'float32': 'numeric',
+        'int32': 'integer',
+        'object': 'text',
+        'bool': 'boolean',
+        'datetime64[ns]': 'timestamp',
+        'datetime64[ns, UTC]': 'timestamp',
+    }
+    return mapping.get(str(dtype), 'text')
+
+
+def _get_geom_col_type(df):
+    geom_col = _get_geom_col_name(df)
+    if geom_col is not None:
+        first_geom = _first_value(df[geom_col])
+        if first_geom:
+            enc_type = detect_encoding_type(first_geom)
+            geom = decode_geometry(first_geom, enc_type)
+            if geom is not None:
+                return geom.geom_type
+        else:
+            warn('Dataset with null geometries')
+
+
+def _first_value(array):
+    array = array.loc[~array.isnull()]  # Remove null values
+    if len(array) > 0:
+        return array.iloc[0]
